refactor(core): merge InputFeature + BlockEditFeature + ArrowNavFeature into KeyboardFeature
