--- packages/core/src/features/arrownav/index.ts
@@ -1,1 +1,0 @@
-export {ArrowNavFeature} from './ArrowNavFeature'
