+++ packages/core/src/features/keyboard/index.ts
@@ -1,0 +1,3 @@
+export {KeyboardFeature} from './KeyboardFeature'
+export {handleBeforeInput, handlePaste, replaceAllContentWith, applySpanInput} from './input'
+export {getCaretRawPosInBlock, getDomRawPos, getDomRawPosInMark, setCaretAtRawPos} from './rawPosition'
